update feeds to use the new api
